fix : common 모듈에서 예외 관리 안함
