+++ common-module/build.gradle
@@ -2,19 +2,5 @@ bootJar { enabled = false }
 jar { enabled = true }
 
 dependencies {
-    //Web
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    //Test
-    testImplementation 'org.springframework.boot:spring-boot-starter-test'
-    //JPA
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-    //QueryDSL
-    implementation 'com.querydsl:querydsl-jpa:5.0.0:jakarta'
-    annotationProcessor "com.querydsl:querydsl-apt:${dependencyManagement.importedProperties['querydsl.version']}:jakarta"
-    annotationProcessor "jakarta.annotation:jakarta.annotation-api"
-    annotationProcessor "jakarta.persistence:jakarta.persistence-api"
-}
 
-clean {
-    delete file('src/main/generated')
 }
