--- common-module/src/main/java/com/lolclone/common_module/authenticationserver/domain/authentication/Authentication.java
@@ -1,9 +1,0 @@
-package com.lolclone.common_module.authenticationserver.domain.authentication;
-
-import com.lolclone.common_module.authenticationserver.Role;
-
-public interface Authentication {
-    Long getId();
-
-    Role getRole();
-}
