--- common-module/src/main/java/com/lolclone/common_module/commonexception/BadRequestException.java
@@ -1,9 +1,0 @@
-package com.lolclone.common_module.commonexception;
-
-import com.lolclone.common_module.exception.domain.ExceptionType;
-
-public class BadRequestException extends FlowOfEssenceException {
-    public BadRequestException(ExceptionType exceptionType) {
-        super(exceptionType);
-    }
-}
