--- common-module/src/main/java/com/lolclone/common_module/commonexception/ForbiddenException.java
@@ -1,9 +1,0 @@
-package com.lolclone.common_module.commonexception;
-
-import com.lolclone.common_module.exception.domain.ExceptionType;
-
-public class ForbiddenException extends FlowOfEssenceException {
-    public ForbiddenException(ExceptionType exceptionType) {
-        super(exceptionType);
-    }
-}
