--- common-module/src/main/java/com/lolclone/common_module/commonexception/UnauthorizedException.java
@@ -1,9 +1,0 @@
-package com.lolclone.common_module.commonexception;
-
-import com.lolclone.common_module.exception.domain.ExceptionType;
-
-public class UnauthorizedException extends FlowOfEssenceException {
-    public UnauthorizedException(ExceptionType exceptionType) {
-        super(exceptionType);
-    }
-}
