+++ common-module/src/main/java/com/lolclone/commonmodule/CommonModuleApplication.java
@@ -1,4 +1,4 @@
-package com.lolclone.common_module;
+package com.lolclone.commonmodule;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class CommonModuleApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(CommonModuleApplication.class, args);
-	}
+    public static void main(String[] args) {
+        SpringApplication.run(CommonModuleApplication.class, args);
+    }
 
 }
