+++ common-module/src/main/java/com/lolclone/commonmodule/apigatewayserver/domain/AuthenticateContext.java
@@ -1,0 +1,7 @@
+package com.lolclone.commonmodule.apigatewayserver.domain;
+
+public interface AuthenticateContext {
+    Long getId();
+    Role getRole();
+    Authentication getAuthentication();
+}
