+++ common-module/src/main/java/com/lolclone/commonmodule/apigatewayserver/domain/Authentication.java
@@ -1,0 +1,7 @@
+package com.lolclone.commonmodule.apigatewayserver.domain;
+
+public interface Authentication {
+    Long getId();
+
+    Role getRole();
+}
