+++ common-module/src/main/java/com/lolclone/commonmodule/exception/UnexpectedException.java
@@ -1,0 +1,10 @@
+package com.lolclone.commonmodule.exception;
+
+public class UnexpectedException extends RuntimeException{
+    private final String code;
+
+    public UnexpectedException(String message, String code) {
+        super(message);
+        this.code = code;
+    }
+}
