--- common-module/src/main/resources/application.yml
@@ -1,1 +1,0 @@
-
