--- common-module/src/test/java/com/lolclone/common_module/CommonModuleApplicationTests.java
@@ -1,13 +1,0 @@
-package com.lolclone.common_module;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CommonModuleApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
